Started Accounts
Implemented the features to get accounts somewhat functional. You can 'sign in' but when you 'sign in' the page doesn't change. The logic is present though so it will check the database if the credentials match. And I created a table to store account information and make sure that the passwords are stored securely